Rename format/parseBalance methods to lif2LifWei form
They convert from Lif units to fraction of Lif units. Same as
ethToWei.  While working on the codebase I found it confusing
to know when to use which of them.
Even though I'm just using LifWei to represent "fraction of a
Lif", we should find a good name for it"