Fix weiRaised calculation: recalculate on every bid
weiRaised is dynamic based on the lastPrice. So it needs to be recalculated on
every bid, as tokensSold * lastPrice. This is because bids with price higher than
lastPrice will have the opportunity to claim the difference between the wei sent and
the wei needed to buy that qty of tokens at the lastPrice