Call calculateDistributionPeriods as the owner, not in constructor
Avoids out of gas exception in the constructor. We need to find a better way
in the future because we should not depend on the owner calling this
function, the market maker should be autonomous as possible. But for now
is ok so we can keep building the rest of the stuff