Make the market maker test scenario to run
Fixes the Market Maker and the test

Uses chai-bignumber to make assertions on eth balance

Learned about bignumber gotchas with web3js in
https://github.com/ethereum/wiki/wiki/JavaScript-API#a-note-on-big-numbers-in-web3js

The test now uses a state object and passes it around. Every command
mutates this state object. This can work as the basis for a future
property-based test :)