Tests: Add isInvalidOpcodeEx and improve exception handling
We were considering any exception in the shouldThrow usual block in the
gen tests commands. The idea is that shouldThrow means whether the
transaction should throw on the blockchain or not, and it's usually
with the invalid opcode error, not with any error. This change makes
it so we only accept invalid opcode exception as the expected thrown
exception. Other kinds of exceptions should make the test to fail

So basically: When it's not invalid opcode, it means there was some other
exception instead of the usual exception from the contract