Do not output testrpc memory usage in CI
Stops sending testrpc memory debug info to stdout in travis. It was used
track memory usage in case of testrpc crashing in CI, but we are not
really using it and those crashes are not that frequent. Instead, we can
enable DEBUG output for the tests in a further commit