Fix time drift in MVM timestamp in tests
nextTimestamp taken from 'latest' block before running tx is not a
robust way of getting the timestamp used during the creation of
the MVM contract. We have to actually get the timestamp after the
block was mined to get a consistent timestamp. This was causing
random build failures that were quite hard to debug